# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Trueke Bogot\u00E1 Solidario API",
    "description": "Errores: application/problem\u002Bjson (title = mensaje mostrable, errors = validaci\u00F3n por campo). Fechas en UTC ISO-8601 con Z. Enums como texto. Notificaciones en tiempo real: SignalR en /hubs/notificaciones (token por ?access_token=, evento \u0022notificacion\u0022).",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/admin/usuarios/{usuarioId}/rol": {
      "patch": {
        "tags": [
          "Administracion"
        ],
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarRolRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarRolRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CambiarRolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/usuarios/{id}/publicaciones": {
      "get": {
        "tags": [
          "Perfiles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Pagina",
            "in": "query",
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Tamano",
            "in": "query",
            "schema": {
              "maximum": 50,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/publicaciones": {
      "get": {
        "tags": [
          "Publicaciones"
        ],
        "parameters": [
          {
            "name": "CategoriaId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Modo",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ModoDto"
            }
          },
          {
            "name": "Localidad",
            "in": "query",
            "schema": {
              "maxLength": 60,
              "type": "string"
            }
          },
          {
            "name": "Texto",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "PrecioMin",
            "in": "query",
            "schema": {
              "maximum": 1000000000,
              "minimum": 0,
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "PrecioMax",
            "in": "query",
            "schema": {
              "maximum": 1000000000,
              "minimum": 0,
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "SoloVerificados",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Orden",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/OrdenPublicacionesDto"
            }
          },
          {
            "name": "Pagina",
            "in": "query",
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Tamano",
            "in": "query",
            "schema": {
              "maximum": 50,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Publicaciones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearPublicacionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearPublicacionRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CrearPublicacionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/publicaciones/cercanas": {
      "get": {
        "tags": [
          "Publicaciones"
        ],
        "parameters": [
          {
            "name": "Lat",
            "in": "query",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "Lon",
            "in": "query",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "RadioKm",
            "in": "query",
            "schema": {
              "maximum": 50,
              "minimum": 0.1,
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "CategoriaId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Modo",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ModoDto"
            }
          },
          {
            "name": "Max",
            "in": "query",
            "schema": {
              "maximum": 50,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicacionCercanaDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicacionCercanaDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicacionCercanaDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/publicaciones/{id}": {
      "get": {
        "tags": [
          "Publicaciones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Publicaciones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearPublicacionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearPublicacionRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CrearPublicacionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/publicaciones/{id}/favorito": {
      "post": {
        "tags": [
          "Publicaciones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Publicaciones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/favoritos": {
      "get": {
        "tags": [
          "Publicaciones"
        ],
        "parameters": [
          {
            "name": "Pagina",
            "in": "query",
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Tamano",
            "in": "query",
            "schema": {
              "maximum": 50,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionDtoPaginaDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/publicaciones/{id}/cancelar": {
      "post": {
        "tags": [
          "Publicaciones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelarPublicacionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelarPublicacionRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CancelarPublicacionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/publicaciones/{id}/destacar-gratis": {
      "post": {
        "tags": [
          "Publicaciones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CambiarRolRequest": {
        "type": "object",
        "properties": {
          "rol": {
            "$ref": "#/components/schemas/RolDto"
          }
        },
        "additionalProperties": false
      },
      "CancelarPublicacionRequest": {
        "type": "object",
        "properties": {
          "motivo": {
            "maxLength": 300,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CategoriaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "CrearPublicacionRequest": {
        "required": [
          "descripcion",
          "localidad",
          "titulo"
        ],
        "type": "object",
        "properties": {
          "titulo": {
            "maxLength": 120,
            "minLength": 3,
            "type": "string"
          },
          "descripcion": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          },
          "categoriaId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "modo": {
            "$ref": "#/components/schemas/ModoDto"
          },
          "localidad": {
            "maxLength": 60,
            "minLength": 2,
            "type": "string"
          },
          "precioReferenciaCop": {
            "maximum": 1000000000,
            "minimum": 0.01,
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "latitud": {
            "maximum": 90,
            "minimum": -90,
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitud": {
            "maximum": 180,
            "minimum": -180,
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "imagenes": {
            "maxItems": 5,
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ModoDto": {
        "enum": [
          "Trueke",
          "Compra",
          "Donacion"
        ],
        "type": "string"
      },
...
      "OrdenPublicacionesDto": {
        "enum": [
          "Recientes",
          "PrecioAsc",
          "PrecioDesc"
        ],
        "type": "string"
      },
...
      "PerfilPublicoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "localidad": {
            "type": "string"
          },
          "reputacion": {
            "type": "number",
            "format": "double"
          },
          "verificado": {
            "type": "boolean"
          },
          "tipoCuenta": {
            "type": "string"
          },
          "calificacionPromedio": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "totalCalificaciones": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PublicacionCercanaDto": {
        "type": "object",
        "properties": {
          "publicacion": {
            "$ref": "#/components/schemas/PublicacionDto"
          },
          "distanciaKm": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "PublicacionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "titulo": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "categoria": {
            "$ref": "#/components/schemas/CategoriaDto"
          },
          "modo": {
            "type": "string"
          },
          "precioReferenciaCop": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "localidad": {
            "type": "string"
          },
          "latitud": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitud": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "coordenadasAproximadas": {
            "type": "boolean"
          },
          "imagenes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estado": {
            "type": "string"
          },
          "fechaPublicacion": {
            "type": "string",
            "format": "date-time"
          },
          "fechaEdicion": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "destacada": {
            "type": "boolean"
          },
          "destacadaHasta": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "propietario": {
            "$ref": "#/components/schemas/PerfilPublicoDto"
          },
          "esMia": {
            "type": "boolean"
          },
          "esFavorita": {
            "type": "boolean"
          },
          "oculta": {
            "type": "boolean"
          },
          "motivoOcultamiento": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PublicacionDtoPaginaDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicacionDto"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "pagina": {
            "type": "integer",
            "format": "int32"
          },
          "tamano": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "RolDto": {
        "enum": [
          "Cliente",
          "Administrador",
          "SuperUsuario"
        ],
        "type": "string"
      },
...
      "UsuarioDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombreCompleto": {
            "type": "string"
          },
          "localidad": {
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "rol": {
            "type": "string"
          },
          "tipoCuenta": {
            "type": "string"
          },
          "planVigenteHasta": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "verificado": {
            "type": "boolean"
          },
          "estadoVerificacion": {
            "type": "string"
          },
          "saldoEcoPuntos": {
            "type": "integer",
            "format": "int32"
          },
          "reputacion": {
            "type": "number",
            "format": "double"
          },
          "truekesCompletados": {
            "type": "integer",
            "format": "int32"
          },
          "comprasRealizadas": {
            "type": "integer",
            "format": "int32"
          },
          "donacionesRealizadas": {
            "type": "integer",
            "format": "int32"
          },
          "destacadosGratisRestantes": {
            "type": "integer",
            "format": "int32"
          },
          "correoVerificado": {
            "type": "boolean"
          },
          "tieneClave": {
            "type": "boolean"
          },
          "vinculadoGoogle": {
            "type": "boolean"
          },
          "dosFactoresActivo": {
            "type": "boolean"
          },
          "codigosRecuperacionRestantes": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Pega el token devuelto por /auth/login (sin la palabra Bearer).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}